Everywhere: Moved wiki to mkdocs
Even though the integrated repository wiki from github is nice, it doesn't allow for managing contributions
through pull requests and issues. Because of that, I've decided to migrate the wiki to mkdocs.

Most of the code came from https://github.com/codeoverflow-org/nodecg-io-docs as they already have
a github actions setup in place as well as do not use google fonts.