Index: Clarify role of mbp-2016-linux guide link
As currently discussed in https://github.com/Dunedan/mbp-2016-linux/pull/171,
the t2linux wiki is more of a user centric approch to documentation. It's still
valuable to have technical documentation, the mbp-2016-linux repository
provides exactly that. For someone looking for such documentation,
it might previously not have been clear that "State of Linux on the MacBook Pro"
refercenes it.