distros: FAQs: add update kernel instructions
These can be refered to throughtout the wiki, without having to repeat
instructions.

I don't know what the method for doing this on Gentoo would be, so if
someone knows what that would be, feel free to add it.

The added instructions will probably be simplified once we get package
repos on (probably) dl.t2linux.org.