arch: install: only use t2linux/archiso-t2 ISO
Before there were a bunch of notes for when using the the
t2linux/archiso-t2 ISO. Made these the default instructions, and removed
instructions for the older ISO. This means users won't have to worry
about getting wired internet.

This includes changing the repo added to my one, but once we get updated
kernels and so forth on the dl.t2linux.org one, that should be changed
back.

Additionally, the archinstall script is very outdated, I don't know
if/when I'll update it, so I've removed the link to it.