Update get-apple-firmware.service activation logic (#624)
This creates the file /etc/get_apple_firmware_attempted when the service
is activated. The service will only activate when this file is *not*
present. The service will run once on first boot and will disable itself
after, regardless of the exit status of the firmware script. This change
should make the service actually start on most distros since many do not
follow systemd first-boot conventions.